\name{nMDNA}
\alias{nMDNA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
nM DNA
}
\description{
Calculates nano moles of dsDNA from the concentration and size
}
\usage{
nMDNA(conc, size)
}
\arguments{
  \item{conc}{
concentration of DNA in solution in ng/uL
}
  \item{size}{
length of dsDNA in base pairs
}
}
\details{

}
\value{
returns the molar concentration of dsDNA in solution in nM
}
\references{

}
\author{
Julian Spagnuolo
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
nMDNA(conc=4.8, size=316)
23.01496

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
}
\keyword{ ~nM }
\keyword{ ~DNA }
\keyword{ ~dna }
\keyword{ ~nm }
